Changes in ThirdPartyServer:
Added the ProcessUserDatabaseUpdate method to handle incoming posts from the authenticator requesting to update the third party database due to changes in the the authenticator database.
Added helper method GetMessageParametersAndEnsureAuthenticatorIsSender which retrieves the message parameters of a request and ensures that the message was send by the authenticator.
Added the new case in the switch in the ProcessIncomingPost method (/request=userdeleted/).
Cleaned up some code (removed some old code that was comment'ed out).