Updated ThirdPartyUserAccount constructor to take another parameter: PkiIdEmail. This was done to associate ThirdPartyUserAccounts with specific PKI identifiers. Also added a property to retrieve the ID of a target account.
Updated AddUserAccount in ThirdParty to reflect the new parameter in ThirdPartyUserAccount constructor.
Added method PkiIdForAccount to ThirdParty to retrieve the PKI id for a specific account in the third party database.
Removed the TestUser in ThirdParty constructor.

ThirdPartyServer had 3 fields changed to readonly.
Updated the methods AnswerAuthTokenPost and AnserUserTokenPost in ThirdParty. This was done to move the parameter-fetching to these methods instead of the ProcessIncomingPost. Also this was changed to allow proper verification of message origins.

ThirdPartyManualTests was updated to reflect the changes to the constructor in ThirdPartyUserAccount.